// src/modules/garmentsProduction/reports/sewing/sewing-input-challan-report/sewing-input-challan-report-type.ts
export type SewingInputChallanReportType = {
  FACTORYID: number;
  BUYERID: number;
  BUYERNAME?: string;
  STYLEID: number;
  STYLENO?: string;
  STYLENAME?: string;
  FLOORID: number;
  FLOORNAME?: string;
  LINEID: number;
  LINENAME?: string;
  SEWINGISSUENO?: string;
  SEWINGINPUTDATE?: string;
  ACTUALETD?: string;
  PONO?: string;
  COLORID: number;
  COLORNAME?: string;
  SIZEID: number;
  SIZENAME?: string;
  SEWINGINPUTQTY: number;
  DESTINATION?: string;
  ISSEDQTY: number;
  REMARKS?: string;
  COMPANYNAME?: string;
  COMPANYADDRESS?: string;
  COMPANYUTILITY?: string;
  SORTINGNO: number;
  BATCH_NO?: string;
  SIZE_SET_STATUS?: string;
  PREPARED_BY?: string;
  PO_QTY: number;
};